Refactor delete methods to use int instead of Long and improve error handling; update UI components for better clarity and functionality

+++ foc/src/main/java/com/alxbryann/foc/persistence/IncomeJpaController.java
@@ -52,7 +52,7 @@ public class IncomeJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) {
+    public void destroy(int id) {
         EntityManager em = getEntityManager();
         EntityTransaction transaction = em.getTransaction();
         try {
